[Model] Qwen4Exp QSA fp8 test: quantize in bf16 to avoid a large fp32 temp
The two largest fp8 cases (tp1_r2048_*) hit a PyTorch CUDACachingAllocator
assert (NVML_SUCCESS == r) during the `.float()` upcast of the whole K/V
cache on the H200 CI. The fp32 intermediate was unnecessary — the scales are
exact powers of two and the fp8 (e4m3) rounding dominates — so quantize
directly in bf16. This avoids the large allocation that triggered the assert
and keeps the stored values identical.

+++ tests/models/qwen4_exp/test_qsa_reference.py
@@ -1045,8 +1045,8 @@ def test_qsa_sparse_paged_attention_correctness(
         # and catches a k/v swap; scales are host floats, as the layer exposes
         # them. Stored values are the scaled ones, as reshape_and_cache does.
         k_scale, v_scale = 0.5, 2.0
-        k_cache = (k_cache.float() / k_scale).to(torch.float8_e4m3fn)
-        v_cache = (v_cache.float() / v_scale).to(torch.float8_e4m3fn)
+        k_cache = (k_cache / k_scale).to(torch.float8_e4m3fn)
+        v_cache = (v_cache / v_scale).to(torch.float8_e4m3fn)
     else:
         k_scale, v_scale = 1.0, 1.0
